item sales chart
good news, i managed to do the chart. bad news: it's not quite right. Take a look at it and you'll notice the sales for UBERMAN hair gel in may is 0, but in the sales history table it's 1. I cant figure out why it's like that. Oh and to view this please download the latest SQL file that i'm about to upload