THE BUG IS FIXED
oh my god doing this chart was so stressful all it took to fix the bug was populate the arrays with 0s and then overwrite it???? i'm probably not making sense but god im glad its over. ALSO for your comfort, I've added navigation.php and included that file to all pages. so now if you want to make changes to the navigation, just edit navigation.php